backend/requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
Flask-Limiter==2.4.0
python-dotenv==0.19.0
bleach==4.1.0
gunicorn==20.1.0
Werkzeug==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
cryptography==3.4.7
requests==2.26.0
python-dateutil==2.8.2
pytz==2021.1
blinker==1.4
itsdangerous==2.0.1
sentry-sdk[flask]==1.14.0
prometheus-flask-exporter==0.22.4
Flask-MonitoringDashboard==3.1.1
psutil==5.9.0 